Fix crash when asking about non existing domains
When asking for non-existing domains/rtypes *in list context*, an undef
value would be used to dereference to an array. This would lead to a
crash with the following error:

> Can't use an undefined value as an ARRAY
>     at lib/Parse/DNS/Zone.pm line 208.

If we were to have asked about this domain before in scalar context,
however, the data structure would have been initialized, and get_rdata
would have been able to return an emtpy list as it should have.